libABC,Uploader: make sure the queue runs again for retries.
This patch address a few shortcomings of the previous patch.
First of all, although UploadItem did emit a signal with progress
equal to -1, it did not update its own value for progress, so the
check in the Uploader would not find that item->progress() < 0, and
would not insert the item in the `retryItems` queue.

Additionally, the queue was not run again after the retry operation
had failed, so the time would keep firing off at doubling intervals,
but do nothing.